Inject inline defs inside lambda functions in defn, let, defmethod and fn.
Two major changes:
I have swapped the definitions of defn* with fn*, and vice versa.
This makes more sense because defn* is just fn* with a name.
This was also done so that I could
convert all fn forms in a form to *fn.

Previously I was only injecting inline defs for a function's arguments and let bindings
Now I'm also injecting it inside lambdas that maybe present in a function.

This addition was primarily inspired by a usecase where a user had a custom let-like macro,
which macroexpanded to a complex form, while the complex form  expanded to include let* and bindings,
the original symbols were not a part of these, rather they were a part of an anonymous function.

```clojure
(mylet [a 1
        b 2]
       (+ a b))
;; expanded to something like this
;; this isn't exact, but you get the essence
(fn [a]
   ((fn [b]
      (+ a b))
    2)
 1) ; 3
```
The reason was this allowed some custom error-handling but that's not important to this.

There were two ways to solve this problem.
1. Inject inline defs inside lambdas,
2. Let users declare macros that were like let and inject inline defs inside them BEFORE calling macroexpand

The reason I didn't go with #2 was that the approach would be harder to implement, users would also need to manage configs
There were instances when I would want variables inside lambdas to be defined, but I would have to specially call *fn on them,
so I thought it would be good to go with #1 and kill two birds with 1 stone